// openapi/models/event-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { EventStatus } from './event-status';

/**
 * 
 * @export
 * @interface EventResponse
 */
export interface EventResponse {
    /**
     * 
     * @type {EventStatus}
     * @memberof EventResponse
     */
    'status': EventStatus;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'eventcode': string;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'accepted': number;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'participantkey': string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'participantlink': string;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'calenderid': number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'clip': number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'participant': number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'eventduration': number;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'mattertitle': string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'schedulername': string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'previewlink': string;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'datestart'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'timezone'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'duration'?: string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'stipulation'?: string;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'stipulationobjectid'?: number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'invited': number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'exhibit': number;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'matterid': number;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    'eventtype': EventResponseEventtypeEnum;
    /**
     * 
     * @type {number}
     * @memberof EventResponse
     */
    'tenantid': number;
}

export const EventResponseEventtypeEnum = {
    Deposition: 'Deposition',
    Hearing: 'Hearing',
    Arbitration: 'Arbitration'
} as const;

export type EventResponseEventtypeEnum = typeof EventResponseEventtypeEnum[keyof typeof EventResponseEventtypeEnum];